# include Hermitian{<:Real} in signature for symmetric blas matmul