This dispatches matvec/matmat with `A::Hermitian{<:BlasReal}` to `BLAS.symv!`/`BLAS.symm!` rather than `generic_matvecmul!`/`generic_matmatmul!`:

```julia
using BenchmarkTools
for n in (10, 100, 1000)
    A = rand(n, n)
    b = rand(n)
    B = rand(n, n)
    H = Hermitian(A)
    @btime $H * $b
    @btime $H * $B
end
```

master:
```julia
1.545 μs (1 allocation: 160 bytes)
1.455 μs (5 allocations: 1.09 KiB)
148.136 μs (1 allocation: 896 bytes)
1.564 ms (8 allocations: 78.53 KiB)
19.717 ms (1 allocation: 7.94 KiB)
1.773 s (8 allocations: 7.63 MiB)
```

PR:
```julia
561.506 ns (1 allocation: 160 bytes)
392.736 ns (1 allocation: 896 bytes)
3.199 μs (1 allocation: 896 bytes)
47.599 μs (2 allocations: 78.20 KiB)
99.609 μs (1 allocation: 7.94 KiB)
31.843 ms (2 allocations: 7.63 MiB)
```

Edit: updated with mat-mat benchmarks too.