@tkelman 2017-06-10
Just leaving a note that these methods wouldn't give the right answers if we allowed nonzero imaginary components on the diagonal in the Hermitian constructor.